refactor(bisect): move basic strategies into `basic` module
